Change the structure of the xml quite a bit with hacks on top of hacks on top of hacks. There is now functionality between the fcl file and this xml generation script to create monte carlo samples which use a generator fcl file as usual but also uses non-standard downstream fcl files. Necessary to make similar MC samples with variation in noise, electron lifetime, to run filters etc.
To use an alternative downstream fcl file, someone has to specify some comments in the generator-level fcl file that this script looks for
Each comment line must start #MCCSTATEMENT
An MCCSTATEMENT line is a comma-separated list of keys and values where the key specifies a production stage and a value specifies a fcl file.
The MCCSTATEMENT is also a key and its value a string appended to the project name (so its distinguishible from the nominal project).
The keys are MCCSTATEMENT: G4: DETSIM: RECO: ANATREE:
An example line might look like
MCCSTATEMENT: cool_project, G4: cool_g4.fcl, DETSIM: super_cool_electronics.fcl
Note that not every key is specified in the example, defaults are assumed in those cases
You can specify as many MCCSTATEMENTS as you like.

There is also a special string searched for which tells the xml generation script if it should NOT make the default project.
MCCSTATEMENT: SKIPDEFAULT
Not specifying the above means the default project xml is also made.

You really shouldn't look at the alterations though.  The hacks are horrible.